requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==2.8
daphne==3.0.2
django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyopenssl==20.0.1
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0